+++ config/config.go
@@ -5,3 +5,9 @@ const CsvURL string = "https://wizeline-go-bootcamp-canodez.s3.amazonaws.com/civ
 const SteamApiKey string = "F26304D6CD0C3ED16903A24351072F56"
 const SteamApiURL string = "http://api.steampowered.com/ISteamUser/ResolveVanityURL/v1/"
 const AOE2ApiLastGame string = "https://aoe2.net/api/player/lastmatch?game=aoe2de&steam_id="
+const AOE2ApiLeaderboard string = "https://aoe2.net/api/leaderboard?game=aoe2de&leaderboard_id=3&start=1&count="
+const ErrMessage string = "There was an error"
+const CsvName string = "result.csv"
+const CsvPath string = "./tmp/" + CsvName
+const ItemsDefault string = "100"
+const Port string = ":8080"
