+++ response/response.go
@@ -7,7 +7,6 @@ import (
 )
 
 func NewResponse(response map[string]string, elapsed time.Time) logrus.Fields {
-	// t := Todo{Result: response, Completed: true, Due: time.Now()}
 
 	start := time.Now()
 
@@ -21,12 +20,24 @@ func NewResponse(response map[string]string, elapsed time.Time) logrus.Fields {
 }
 
 func NewResponseIndex(response map[int][]string, elapsed time.Time) logrus.Fields {
-	// t := Todo{Result: response, Completed: true, Due: time.Now()}
 
 	start := time.Now()
 
 	fields := logrus.Fields{
 		"Response":  response,
+		"Completed": true,
+		"Time":      start.Sub(elapsed).Milliseconds(),
+	}
+
+	return fields
+}
+
+func NewResponseInterface(response map[string]interface{}, elapsed time.Time) logrus.Fields {
+
+	start := time.Now()
+
+	fields := logrus.Fields{
+		"Body":      response,
 		"Completed": true,
 		"Time":      start.Sub(elapsed).Milliseconds(),
 	}
