+++ router/router.go
@@ -3,6 +3,7 @@ package router
 import (
 	"net/http"
 
+	"Project/config"
 	"Project/handler"
 
 	"github.com/gorilla/mux"
@@ -12,12 +13,10 @@ import (
 func HandleRequests() {
 	router := mux.NewRouter().StrictSlash(true)
 
-	router.HandleFunc("/civ/{id}/", handler.GetCivilizationById)
-	router.HandleFunc("/lastGame/{user}", handler.GetLastGameData)
-	router.HandleFunc("/leaderboard", handler.GetLeaderboard100)
+	router.HandleFunc("/leaderboard", handler.GetLeaderboardDefault)
 	router.HandleFunc("/leaderboard/{count}", handler.GetLeaderboardCount)
 	router.HandleFunc("/leaderbycountry/{country}", handler.GetLeaderByCountry)
-	router.HandleFunc("/read/{type}/{items}/{items_per_workers}", handler.ReadItems)
+	router.HandleFunc("/read/{type}/{items}/{itemsPerWorker}", handler.ReadItems)
 
-	log.Error(http.ListenAndServe(":80", router))
+	log.Error(http.ListenAndServe(config.Port, router))
 }
