+++ utils/file.go
@@ -1,47 +1,13 @@
 package utils
 
 import (
-	"Project/config"
 	"Project/model"
 	"encoding/csv"
-	"io"
-	"net/http"
 	"os"
 	"strconv"
 
 	log "github.com/sirupsen/logrus"
 )
-
-func DownloadFile(filepath string, url string) error {
-
-	if _, err := os.Stat("./tmp/" + config.CsvFileName); err == nil {
-		log.Debug("File already exist")
-		return nil
-
-	} else {
-
-		log.Debug("File didn't exist")
-
-		// Get the data
-		resp, err := http.Get(url)
-		if err != nil {
-			return err
-		}
-		defer resp.Body.Close()
-
-		// Create the file
-		out, err := os.Create(filepath)
-		if err != nil {
-			return err
-		}
-		defer out.Close()
-
-		// Write the body to file
-		_, err = io.Copy(out, resp.Body)
-		return err
-	}
-
-}
 
 func CreateCSV(data []model.Player) {
 
@@ -77,8 +43,4 @@ func checkError(message string, err error) {
 	if err != nil {
 		log.Fatal(message, err)
 	}
-}
-
-func ReadCSV() {
-
 }
